Split JS into server and API.
- Helpers in serverHelpers.js
